AI-based Content Generation System/Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq==0.12.0
replicate==1.0.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.3
replicate==1.0.4
pytest==7.3.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0